+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+- Add `Enumerables.ZipEqual()`, an alternative to LINQ's `Zip()`
+  that ensures both sequences have the same length.
+
 ## [1.24.0] - 2017-05-29
 - Add `SmartList<T>` class, which wraps an enumerable to provide a list interface,
   while minimizing enumerating the enumerable and ensuring it is enumerated only once.